- Fix warning
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_1@208160 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/reflection/php_reflection.c
@@ -554,7 +554,7 @@ static zend_op* _get_recv_op(zend_op_array *op_array, zend_uint offset)
 	++offset;
 	while (op < end) {
 		if ((op->opcode == ZEND_RECV || op->opcode == ZEND_RECV_INIT) &&
-		    op->op1.u.constant.value.lval == offset) {
+		    op->op1.u.constant.value.lval == (long)offset) {
 		  return op;
 	  }
 	  ++op;
